fix #1446
