+++ scripts.go
@@ -1,0 +1,29 @@
+package main
+
+import (
+	"bytes"
+	"io/ioutil"
+
+	"github.com/PuerkitoBio/goquery"
+	"github.com/russross/blackfriday"
+)
+
+func readme() []byte {
+	input, err := ioutil.ReadFile("./README.md")
+	if err != nil {
+		panic(err)
+	}
+	html := append([]byte("<body>"), blackfriday.MarkdownCommon(input)...)
+	html = append(html, []byte("</body>")...)
+	return html
+}
+
+func startQuery() *goquery.Document {
+	buf := bytes.NewBuffer(readme())
+	query, err := goquery.NewDocumentFromReader(buf)
+	if err != nil {
+		panic(err)
+	}
+
+	return query
+}
